refactor(discovery): split plugin.ts, validate limit, add regression tests
- Split monolithic src/plugin.ts (626 lines) into focused modules under the
  500-line gate: types.ts, utils.ts, cache.ts, capabilities.ts, discovery.ts,
  entry plugin.ts (107 lines). Behavior unchanged; entry still exports only
  NineRouterPlugin.
- Harden mapRouterCapabilities: limit.context/output now pass through
  toPositiveInt (rejects 0/negative/NaN/Infinity/float/junk-string), nested
  capabilities.* take precedence over top-level context_length/
  max_completion_tokens fallback.
- Restore models.dev limit passthrough in findModelsDevMatch by mirroring
  bestMatch.limit onto capabilities.contextWindow/maxOutput so catalog
  limits reach config.limit via the single validated path.
- Add test-limit.mjs regression suite (limit populated from nested, top-level,
  string-coerced, and models.dev catalog paths; invalid values never produce
  a limit) and npm test script (pretest build + both suites).
